+++ src/lib/caminante/caminante.hpp
@@ -34,7 +34,7 @@ class Caminante
 
 private:
     int camino[400];
-    float fitness;
+    double fitness;
 public:
 
     Caminante();
@@ -55,7 +55,7 @@ public:
     void calcMiFit(int **dist, int numCiuds);
 
     //Devuelve el fitness del caminante.
-    float MyFit();
+    double MyFit();
 
     //Función de mutar.
     void mutar(const int numCities);
@@ -104,7 +104,7 @@ public:
     //Devuelve el porcentaje de caminantes que son mejores que el fit que le introducimos,
     //tambien por mejorFit devuelve el fitness del mejor caminante y por media la
     //media de fitness de los caminates
-    float stats(float fit, float &mejorFit, float &media);
+    double stats(double fit, double &mejorFit, double &media);
     //divide la poblacion en n subpoblaciones
     void dividir(int n, Poblacion pobs[]);
